actions/src/py_soundfile/requirements.txt
soundfile
numpy
python-swiftclient
